+++ src/gui/view/asyncinflater.cc
@@ -17,6 +17,7 @@
  *********************************************************************************/
 #include <thread>
 #include <view/asyncinflater.h>
+#include <porting/cdlog.h>
 namespace cdroid{
 
 AsyncLayoutInflater::AsyncLayoutInflater(Context* context) {
@@ -111,7 +112,7 @@ void AsyncLayoutInflater::InflateThread::runInner() {
         LOGW("%s",ex.what());
         return;
     }
-LOGD("request=%p resid=%s",request,request->resid.c_str());
+
     try {
         request->view = request->inflater->mInflater->inflate(
                 request->resid, request->parent, false);